feat: separate results to server and client
