+++ src/app/(client)/_components/data/results-interactive.tsx
@@ -1,19 +1,21 @@
 'use client';
 import { useRouter, useParams, useSearchParams } from 'next/navigation';
+import { useSelector } from 'react-redux';
 
-import { cn } from '@/app/(server)/_lib/cn';
 import type { Character } from '@/app/_types/character';
 
-import { CharacterCard } from './character-card';
+import { CharacterCard } from '../../../(server)/_components/data/character-card';
+import type { RootState } from '../../../_store/store';
 
 type ResultsProps = {
   characters: Character[];
 };
 
-function Results({ characters }: ResultsProps) {
+function ResultsInteractive({ characters }: ResultsProps) {
   const router = useRouter();
   const { id } = useParams();
   const searchParams = useSearchParams();
+  const selectedIds = useSelector((state: RootState) => state.selectedCharacters.ids);
 
   const handleCharacterClick = (characterId: number) => {
     const params = new URLSearchParams(searchParams);
@@ -21,14 +23,7 @@ function Results({ characters }: ResultsProps) {
   };
 
   return (
-    <div
-      className={cn(
-        'mx-auto mt-6 flex w-full max-w-6xl',
-        'flex-col items-center rounded-xl border-4 p-4',
-        'shadow-glow',
-      )}
-    >
-      <h2>Characters</h2>
+    <div className="w-full">
       <div className="grid grid-cols-1 gap-4 md:grid-cols-2 lg:grid-cols-4">
         {characters.map((character) => (
           <CharacterCard
@@ -36,6 +31,7 @@ function Results({ characters }: ResultsProps) {
             character={character}
             isActive={id === String(character.id)}
             onClick={() => handleCharacterClick(character.id)}
+            isChecked={selectedIds.includes(character.id)}
           />
         ))}
       </div>
@@ -43,4 +39,4 @@ function Results({ characters }: ResultsProps) {
   );
 }
 
-export default Results;
+export default ResultsInteractive;
