+++ src/app/(server)/_components/data/results.tsx
@@ -1,0 +1,24 @@
+import ResultsInteractive from '@/app/(client)/_components/data/results-interactive';
+import { cn } from '@/app/(server)/_lib/cn';
+import type { Character } from '@/app/_types/character';
+
+type ResultsProps = {
+  characters: Character[];
+};
+
+function Results({ characters }: ResultsProps) {
+  return (
+    <div
+      className={cn(
+        'mx-auto mt-6 flex w-full max-w-6xl',
+        'flex-col items-center rounded-xl border-4 p-4',
+        'shadow-glow',
+      )}
+    >
+      <h2>Characters</h2>
+      <ResultsInteractive characters={characters} />
+    </div>
+  );
+}
+
+export default Results;
